fix wrong type

+++ z80sim/srcsim/iosim.c
@@ -173,7 +173,7 @@ static BYTE p000_in(void)
  */
 static BYTE p001_in(void)
 {
-	extern BYTE getkey(void);
+	extern int getkey(void);
 
 	return ((BYTE) getkey());
 }
